fix: reorganise files
