+++ interapp-frontend/package.json
@@ -41,6 +41,7 @@
     "bun-types": "^1.0.11",
     "eslint": "^8",
     "eslint-config-next": ">=14.1.0",
+    "eslint-import-resolver-custom-alias": "^1.3.2",
     "prettier": "^3.1.0",
     "typescript": "^5.3.3"
   }
