+++ interapp-frontend/src/app/announcements/[id]/edit/EditForm.tsx
@@ -10,7 +10,7 @@ import { useContext, useEffect, useState, useMemo, useCallback } from 'react';
 import { AuthContext } from '@providers/AuthProvider/AuthProvider';
 import { type AnnouncementWithMeta, type AnnouncementForm } from '../../types';
 import { useParams, useRouter } from 'next/navigation';
-import { parseErrorMessage, remapAssetUrl } from '@api/utils';
+import { parseErrorMessage, remapAssetUrl } from '@utils/.';
 import { notifications } from '@mantine/notifications';
 import { Button, Group, TextInput, Title, Text, Stack } from '@mantine/core';
 import { IconClock, IconUser } from '@tabler/icons-react';
