+++ interapp-frontend/src/components/Footer/Footer.tsx
@@ -1,11 +1,18 @@
-import { memo } from 'react';
+import { memo,  } from 'react';
 import { Stack, Text } from '@mantine/core';
 import './styles.css';
 
 const year = new Date().getFullYear();
-const version = process.env.NEXT_PUBLIC_APP_VERSION ? process.env.NEXT_PUBLIC_APP_VERSION : undefined; // check if version is a blank string
 
-const footer = () => {
+const footer = async () => {
+  const getVersion = async () => {
+    "use server";
+    return process.env.NEXT_PUBLIC_APP_VERSION;
+  };
+
+  const version = await getVersion();
+  
+
   return (
     <footer>
       <Stack className='footer' align='center' gap={2}>
