+++ interapp-frontend/tsconfig.json
@@ -23,7 +23,8 @@
       "@api/*": ["./src/api/*"],
       "@providers/*": ["./src/providers/*"],
       "@components/*": ["./src/components/*"],
-      "@hooks/*": ["./src/hooks/*"]
+      "@hooks/*": ["./src/hooks/*"],
+      "@utils/*": ["./src/utils/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
